Actualizacion general y funcional de Preventa para CHINAS

+++ src/main/java/inteldev/android/negocios/BorrarArchivos.java
@@ -1,20 +1,24 @@
 package inteldev.android.negocios;
 
-import android.os.Environment;
-
 import java.io.File;
 
 /**
  * Created by Operador on 6/05/14.
  */
-public class BorrarArchivos {
-    public void BorrarPorContenido(String ubicacion, String contenidoCadena) {
+public class BorrarArchivos
+{
+    public void BorrarPorContenido(String ubicacion, String contenidoCadena)
+    {
         File carpeta = new File(ubicacion);
-
-        for (File archivo : carpeta.listFiles()) {
-            if (archivo.getName().contains(contenidoCadena)) {
-                boolean borro;
-                borro = archivo.delete();
+        if (carpeta.canRead())
+        {
+            for (File archivo : carpeta.listFiles())
+            {
+                if (archivo.getName().contains(contenidoCadena))
+                {
+                    boolean borro;
+                    borro = archivo.delete();
+                }
             }
         }
     }
